requirements.txt
pytest~=7.1.3
selene>=2.0.b8
allure-pytest
python-dotenv==0.20.0
Appium-Python-Client~=2.6.1
selenium~=4.2.0
requests~=2.28.1

pydantic~=1.10.2
Faker~=15.0.0
curlify~=2.2.1
pytest-voluptuous
voluptuous~=0.13.1